feat: enable partial static linking
This removes some libstdc++ dependencies and allows the addon to be used on older systems as well.

The minimum required GLIBC Version is 2.36, which was released August 2022.

Due to it being older than a year, I think it's reasonable to expect that most Vencord users have a system with GLIBC >= 2.36 - and if not they should use the flatpak.